Housekeeping after restructuring
