+++ app/(user)/layout.tsx
@@ -1,4 +1,3 @@
-"use client";
 import Footer from "@/components/Footer";
 import Navbar from "@/components/Navbar";
 import { Analytics } from "@vercel/analytics/react";
