+++ app/(user)/sewa/page.tsx
@@ -1,17 +1,17 @@
+import { simplifiedRental } from "@/app/interface";
+import { client } from "@/app/lib/sanity";
 import { formatCurrency } from "@/utils";
 import Image from "next/image";
 import Link from "next/link";
-import { simplifiedProject } from "../interface";
-import { client } from "../lib/sanity";
 
 async function getData() {
-  const query = `*[_type == "project" && (category->name == "projek baru" || category->name == "subsale")] | order(_createdAt desc) {
+  const query = `*[_type == "rental"] | order(_createdAt desc) {
     _id,
-      "imageUrl": images[0].asset->url,
-      price,
+      installment,
       name,
+      "location": location->name,
       "slug": slug.current,
-      "categoryName": category->name
+      "imageUrl": images[0].asset->url
   }`;
 
   const data = await client.fetch(query);
@@ -22,18 +22,18 @@ async function getData() {
 // Opt out of caching for all data requests in the route segment
 export const dynamic = "force-dynamic";
 
-export default async function BuyList({
+export default async function RentList({
   params,
 }: {
   params: { category: string };
 }) {
-  const data: simplifiedProject[] = await getData();
+  const data: simplifiedRental[] = await getData();
   return (
     <div className="mx-auto max-w-2xl px-6 py-16 sm:px-6 sm:py-24 lg:max-w-7xl lg:px-8">
       <div className="pt-10 flex justify-between items-center">
         <div className="flex-col mx-auto text-center px-2">
           <h1 className="font-playfair mb-4 font-bold uppercase text-4xl md:text-5xl">
-            Beli
+            Sewa
           </h1>
           <a
             href="https://api.whatsapp.com/send?phone=60176515784"
@@ -41,20 +41,20 @@ export default async function BuyList({
             target="_blank"
             className="font-inter text-base md:text-lg hover:font-medium"
           >
-            Hubungi kami jika anda ingin menjual hartanah anda &gt;
+            Hubungi kami jika anda ingin sewakan hartanah anda &gt;
           </a>
         </div>
       </div>
       <div className="mt-12 md:mt-20 grid grid-cols-1 gap-x-6 gap-y-10 sm:grid-cols-2 lg:grid-cols-4 xl:gap-x-8">
-        {data.map((project) => (
+        {data.map((rental) => (
           <div
-            key={project._id}
+            key={rental._id}
             className="group relative border rounded-md shadow-lg hover:shadow-xl"
           >
-            <Link href={`/beli/${project.slug}`}>
+            <Link href={`/sewa/${rental.slug}`}>
               <div className="aspect-square w-full overflow-hidden rounded-t-md lg:h-80">
                 <Image
-                  src={project.imageUrl}
+                  src={rental.imageUrl}
                   alt=""
                   className="w-full h-full object-cover object-center lg:h-full lg:w-full"
                   width={300}
@@ -63,13 +63,13 @@ export default async function BuyList({
               </div>
               <div className="font-inter my-4 mx-4 flex flex-col text-center gap-y-1">
                 <div>
-                  <h2 className="md:text-sm">{project.name}</h2>
+                  <h2 className="md:text-sm">{rental.name}</h2>
                   <span className="absolute left-0 top-3 rounded-r-lg bg-teal-700 px-3 py-1.5 text-xs font-semibold uppercase tracking-wider text-white">
-                    {project.categoryName}
+                    {rental.location}
                   </span>
                 </div>
                 <p className="md:text-sm font-medium">
-                  {formatCurrency(project.price)}
+                  {formatCurrency(rental.installment)}/bulan
                 </p>
               </div>
             </Link>
