--- app/sewa/head.tsx
@@ -1,13 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>Fikry Zulkifly</title>
-      <meta content="width=device-width, initial-scale=1" name="viewport" />
-      <meta name="description" content="Generated by create next app" />
-      <link
-        rel="icon"
-        href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>👋</text></svg>"
-      />
-    </>
-  );
-}
