--- app/sewa/layout.tsx
@@ -1,50 +1,0 @@
-import Footer from "@/components/Footer";
-import Navbar from "@/components/Navbar";
-import { Analytics } from "@vercel/analytics/react";
-import { Inter, Playfair_Display } from "next/font/google";
-import "../../styles/globals.css";
-
-const playfair_display = Playfair_Display({
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-playfair-display",
-});
-
-const inter = Inter({
-  subsets: ["latin"],
-  display: "swap",
-  variable: "--font-inter",
-});
-
-export const metadata = {
-  title: "Next.js",
-  description: "Generated by Next.js",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html
-      lang="en"
-      className={`${playfair_display.variable} ${inter.variable}`}
-    >
-      {/*
-        <head /> will contain the components returned by the nearest parent
-        head.tsx. Find out more at https://beta.nextjs.org/docs/api-reference/file-conventions/head
-      */}
-      <head />
-      {/* <body className="dark:bg-stone-900"> */}
-      <body>
-        {/* <ThemeProvider enableSystem={true} attribute="class"> */}
-        <Navbar />
-        {children}
-        <Footer />
-        {/* </ThemeProvider> */}
-        <Analytics />
-      </body>
-    </html>
-  );
-}
