+++ components/CalculatorSection.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { InitialFormState } from "@/types";
 import { formatCurrency, toNumber } from "@/utils";
-import Link from "next/link";
 import { ChangeEvent, FocusEvent, useState } from "react";
 import ButtonCalculator from "./ButtonCalculator";
 import SlideUp from "./SlideUp";
